// client/src/game/shared.ts
export const DIRECTIONS = ['up', 'down', 'left', 'right'] as const;
export type Direction = typeof DIRECTIONS[number];

export const MOVE_DIRECTIONS = ['m_up', 'm_down', 'm_left', 'm_right'] as const;
export type MoveDirection = typeof MOVE_DIRECTIONS[number];

export const IDLE_DIRECTIONS = ['i_up', 'i_down', 'i_left', 'i_right'] as const;
export type IdleDirection = typeof IDLE_DIRECTIONS[number];

export const SLASH_DIRECTIONS = [
  'sl_up',
  'sl_down',
  'sl_left',
  'sl_right',
] as const;
export type SlashDirection = typeof SLASH_DIRECTIONS[number];

export const THRUST_DIRECTIONS = [
  't_up',
  't_down',
  't_left',
  't_right',
] as const;
export type ThrustDirection = typeof THRUST_DIRECTIONS[number];

export const SPELL_DIRECTIONS = [
  'sp_up',
  'sp_down',
  'sp_left',
  'sp_right',
] as const;
export type SpellDirection = typeof SPELL_DIRECTIONS[number];

export const SHOOT_DIRECTIONS = [
  'sh_up',
  'sh_down',
  'sh_left',
  'sh_right',
] as const;
export type ShootDirection = typeof SHOOT_DIRECTIONS[number];

export const OVERSIZE_DIRECTIONS = [
  'ov_up',
  'ov_down',
  'ov_left',
  'ov_right',
] as const;
export type OversizeDirection = typeof OVERSIZE_DIRECTIONS[number];